#134 refactor: FE 공부방 페이지 - 렌더링 최적화2 - 하단메뉴바 분리 및 state 위치 이동
